chore: move decrypt lit action into lit dir
